problem3_multiagent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2
python-dotenv==1.0.0
requests==2.32.0
streamlit==1.38.0

langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-google-genai==2.0.5

numpy==2.1.0

pdfplumber==0.11.4
diskcache==5.6.3
httpx==0.27.0

setuptools>=68.0.0
wheel